+++ zallet-core/src/commands/repair.rs
@@ -1,1 +1,2 @@
+mod check_witnesses;
 mod truncate_wallet;